move search bar styles back into component
